Move temporal range functions to range feature
