+++ crates/polars-plan/src/dsl/function_expr/range/mod.rs
@@ -1,29 +1,99 @@
+mod date_range;
 mod int_range;
+mod time_range;
+mod utils;
 
 use std::fmt::{Display, Formatter};
 
 use polars_core::prelude::*;
 use polars_core::series::Series;
+use polars_time::{ClosedWindow, Duration};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
+use crate::dsl::function_expr::FieldsMapper;
 use crate::dsl::SpecialEq;
 use crate::map_as_slice;
 use crate::prelude::SeriesUdf;
 
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Clone, Copy, PartialEq, Debug, Eq, Hash)]
+#[derive(Clone, PartialEq, Debug, Eq, Hash)]
 pub enum RangeFunction {
-    IntRange { step: i64 },
-    IntRanges { step: i64 },
+    IntRange {
+        step: i64,
+    },
+    IntRanges {
+        step: i64,
+    },
+    DateRange {
+        every: Duration,
+        closed: ClosedWindow,
+        time_unit: Option<TimeUnit>,
+        time_zone: Option<TimeZone>,
+    },
+    DateRanges {
+        every: Duration,
+        closed: ClosedWindow,
+        time_unit: Option<TimeUnit>,
+        time_zone: Option<TimeZone>,
+    },
+    TimeRange {
+        every: Duration,
+        closed: ClosedWindow,
+    },
+    TimeRanges {
+        every: Duration,
+        closed: ClosedWindow,
+    },
 }
 
 impl RangeFunction {
-    pub(super) fn get_field(&self) -> PolarsResult<Field> {
+    pub(super) fn get_field(&self, mapper: FieldsMapper) -> PolarsResult<Field> {
         use RangeFunction::*;
         let field = match self {
             IntRange { .. } => Field::new("int", DataType::Int64),
             IntRanges { .. } => Field::new("int_range", DataType::List(Box::new(DataType::Int64))),
+            DateRange {
+                every,
+                closed: _,
+                time_unit,
+                time_zone,
+            } => {
+                // output dtype may change based on `every`, `time_unit`, and `time_zone`
+                let dtype = mapper.map_to_date_range_dtype(
+                    every,
+                    time_unit.as_ref(),
+                    time_zone.as_deref(),
+                )?;
+                return Ok(Field::new("date", dtype));
+            },
+            DateRanges {
+                every,
+                closed: _,
+                time_unit,
+                time_zone,
+            } => {
+                // output dtype may change based on `every`, `time_unit`, and `time_zone`
+                let inner_dtype = mapper.map_to_date_range_dtype(
+                    every,
+                    time_unit.as_ref(),
+                    time_zone.as_deref(),
+                )?;
+                return Ok(Field::new(
+                    "date_range",
+                    DataType::List(Box::new(inner_dtype)),
+                ));
+            },
+
+            TimeRange { .. } => {
+                return Ok(Field::new("time", DataType::Time));
+            },
+            TimeRanges { .. } => {
+                return Ok(Field::new(
+                    "time_range",
+                    DataType::List(Box::new(DataType::Time)),
+                ));
+            },
         };
         Ok(field)
     }
@@ -35,6 +105,10 @@ impl Display for RangeFunction {
         let s = match self {
             IntRange { .. } => "int_range",
             IntRanges { .. } => "int_ranges",
+            DateRange { .. } => "date_range",
+            DateRanges { .. } => "date_ranges",
+            TimeRange { .. } => "time_range",
+            TimeRanges { .. } => "time_ranges",
         };
         write!(f, "{s}")
     }
@@ -49,6 +123,40 @@ impl From<RangeFunction> for SpecialEq<Arc<dyn SeriesUdf>> {
             },
             IntRanges { step } => {
                 map_as_slice!(int_range::int_ranges, step)
+            },
+            DateRange {
+                every,
+                closed,
+                time_unit,
+                time_zone,
+            } => {
+                map_as_slice!(
+                    date_range::temporal_range,
+                    every,
+                    closed,
+                    time_unit,
+                    time_zone.clone()
+                )
+            },
+            DateRanges {
+                every,
+                closed,
+                time_unit,
+                time_zone,
+            } => {
+                map_as_slice!(
+                    date_range::temporal_ranges,
+                    every,
+                    closed,
+                    time_unit,
+                    time_zone.clone()
+                )
+            },
+            TimeRange { every, closed } => {
+                map_as_slice!(time_range::time_range, every, closed)
+            },
+            TimeRanges { every, closed } => {
+                map_as_slice!(time_range::time_ranges, every, closed)
             },
         }
     }
