+++ crates/polars-plan/src/dsl/functions/range.rs
@@ -38,33 +38,6 @@ pub fn int_ranges(start: Expr, end: Expr, step: i64) -> Expr {
     }
 }
 
-pub trait Range<T> {
-    fn into_range(self, high: T) -> Expr;
-}
-
-macro_rules! impl_into_range {
-    ($dt: ty) => {
-        impl Range<$dt> for $dt {
-            fn into_range(self, high: $dt) -> Expr {
-                Expr::Literal(LiteralValue::Range {
-                    low: self as i64,
-                    high: high as i64,
-                    data_type: DataType::Int32,
-                })
-            }
-        }
-    };
-}
-
-impl_into_range!(i32);
-impl_into_range!(i64);
-impl_into_range!(u32);
-
-/// Create a range literal.
-pub fn range<T: Range<T>>(low: T, high: T) -> Expr {
-    low.into_range(high)
-}
-
 /// Create a date range from a `start` and `stop` expression.
 #[cfg(feature = "temporal")]
 pub fn date_range(
@@ -79,7 +52,7 @@ pub fn date_range(
 
     Expr::Function {
         input,
-        function: FunctionExpr::TemporalExpr(TemporalFunction::DateRange {
+        function: FunctionExpr::Range(RangeFunction::DateRange {
             every,
             closed,
             time_unit,
@@ -108,7 +81,7 @@ pub fn date_ranges(
 
     Expr::Function {
         input,
-        function: FunctionExpr::TemporalExpr(TemporalFunction::DateRanges {
+        function: FunctionExpr::Range(RangeFunction::DateRanges {
             every,
             closed,
             time_unit,
@@ -130,7 +103,7 @@ pub fn time_range(start: Expr, end: Expr, every: Duration, closed: ClosedWindow)
 
     Expr::Function {
         input,
-        function: FunctionExpr::TemporalExpr(TemporalFunction::TimeRange { every, closed }),
+        function: FunctionExpr::Range(RangeFunction::TimeRange { every, closed }),
         options: FunctionOptions {
             collect_groups: ApplyOptions::ApplyGroups,
             allow_rename: true,
@@ -146,7 +119,7 @@ pub fn time_ranges(start: Expr, end: Expr, every: Duration, closed: ClosedWindow
 
     Expr::Function {
         input,
-        function: FunctionExpr::TemporalExpr(TemporalFunction::TimeRanges { every, closed }),
+        function: FunctionExpr::Range(RangeFunction::TimeRanges { every, closed }),
         options: FunctionOptions {
             collect_groups: ApplyOptions::ApplyGroups,
             allow_rename: true,
@@ -171,4 +144,31 @@ pub fn repeat<E: Into<Expr>>(value: E, n: Expr) -> Expr {
         Ok(Some(s.new_from_index(0, n)))
     };
     apply_binary(value.into(), n, function, GetOutput::same_type()).alias("repeat")
+}
+
+pub trait Range<T> {
+    fn into_range(self, high: T) -> Expr;
+}
+
+macro_rules! impl_into_range {
+    ($dt: ty) => {
+        impl Range<$dt> for $dt {
+            fn into_range(self, high: $dt) -> Expr {
+                Expr::Literal(LiteralValue::Range {
+                    low: self as i64,
+                    high: high as i64,
+                    data_type: DataType::Int32,
+                })
+            }
+        }
+    };
+}
+
+impl_into_range!(i32);
+impl_into_range!(i64);
+impl_into_range!(u32);
+
+/// Create a range literal.
+pub fn range<T: Range<T>>(low: T, high: T) -> Expr {
+    low.into_range(high)
 }
